Fix entrypoint scripts and cache paths on some environments
Modify script modules such that functionality is wrapped in a function
call. This prevents unexpected behavior when importing a module. This
also fixes an issue I had installing and running these entrypoints on
a different environment than I had previously tested on.

Also move the cache path from the location of the onshape_api module to
the user's home directory. This handles environments where the package
is installed globally, in which case the previous cache directory
location is not generally writeable by the user.